Rework and optimisation of initial modification
Complete overhaul on initial modification from scratch. The program would consistently get stuck in loops the cubic length of the initial input, creating massive delays in processing and horrendous overuse of cpu time. The program now correctly builds upon the existing code and is much more optimal, only requiring a few hours over the 3 weeks it would try to run over.

Still a few bugs here and there when switching directories but these are far and few between, and seem to only occur when the program encounters unexpected directory branches.